+++ openpyxl2/descriptors/excel.py
@@ -1,12 +1,12 @@
 from __future__ import absolute_import
-#copyright openpyxl 2010-2015
+#copyright openpyxl 2010-2018
 
 """
 Excel specific descriptors
 """
 
 from openpyxl2.xml.constants import REL_NS
-from openpyxl2.compat import safe_string
+from openpyxl2.compat import safe_string, unicode
 from openpyxl2.xml.functions import Element
 
 from . import (
@@ -44,9 +44,17 @@ class TextPoint(MinMax):
 Coordinate = Integer
 
 
-class Percentage(MatchPattern):
+class Percentage(MinMax):
 
-    pattern = r"((100)|([0-9][0-9]?))(\.[0-9][0-9]?)?%"
+    pattern = r"((100)|([0-9][0-9]?))(\.[0-9][0-9]?)?%" # strict
+    min = -1000000
+    max = 1000000
+
+    def __set__(self, instance, value):
+        if isinstance(value, unicode) and "%" in value:
+            value = value.replace("%", "")
+            value = int(float(value) * 1000)
+        super(Percentage, self).__set__(instance, value)
 
 
 class Extension(Serialisable):