+++ openpyxl2/utils/dataframe.py
@@ -1,14 +1,19 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2018 openpyxl
 
-import numpy
-from pandas import Timestamp
+import operator
+from openpyxl2.compat import accumulate, zip
 
 
 def dataframe_to_rows(df, index=True, header=True):
     """
-    Convert a Pandas dataframe into something suitable for passing into a worksheet
+    Convert a Pandas dataframe into something suitable for passing into a worksheet.
+    If index is True then the index will be included, starting one row below the header.
+    If header is True then column headers will be included starting one column to the right.
+    Formatting should be done by client code.
     """
+    import numpy
+    from pandas import Timestamp
     blocks = df._data.blocks
     ncols = sum(b.shape[0] for b in blocks)
     data = [None] * ncols
@@ -26,10 +31,52 @@ def dataframe_to_rows(df, index=True, header=True):
             data[col_loc] = col
 
     if header:
-        values = list(df.columns.values)
-        if df.columns.dtype.type == numpy.datetime64:
-            values = [Timestamp(v) for v in values]
-        yield [None]*index + values
+        if df.columns.nlevels > 1:
+            rows = expand_levels(df.columns.levels)
+        else:
+            rows = [list(df.columns.values)]
+        for row in rows:
+            n = []
+            for v in row:
+                if isinstance(v, numpy.datetime64):
+                    v = Timestamp(v)
+                n.append(v)
+            row = n
+            if index:
+                row = [None]*df.index.nlevels + row
+            yield row
+
+    cols = None
+    if df.index.nlevels > 1:
+        cols = zip(*expand_levels(df.index.levels))
 
     for idx, v in enumerate(df.index):
-        yield [v]*index + [data[j][idx] for j in range(ncols)]
+        row = [data[j][idx] for j in range(ncols)]
+        if index:
+            yield df.index.names
+            if cols:
+                v = list(next(cols))
+            else:
+                v = [v]
+            row = v + row
+        yield row
+
+
+def expand_levels(levels):
+    """
+    Multiindexes need expanding so that subtitles repeat
+    """
+    widths = (len(s) for s in levels)
+    widths = list(accumulate(widths, operator.mul))
+    size = max(widths)
+
+    for level, width in zip(levels, widths):
+        padding = int(size/width) # how wide a title should be
+        repeat = int(width/len(level)) # how often a title is repeated
+        row = []
+        for v in level:
+            title = [None]*padding
+            title[0] = v
+            row.extend(title)
+        row = row*repeat
+        yield row