+++ openpyxl2/chart/legend.py
@@ -26,7 +26,7 @@ class LegendEntry(Serialisable):
     txPr = Typed(expected_type=RichText, allow_none=True)
     extLst = Typed(expected_type=ExtensionList, allow_none=True)
 
-    __elements__ = ('idx',)
+    __elements__ = ('idx', 'delete', 'txPr')
 
     def __init__(self,
                  idx=0,