+++ doc/changes.rst
@@ -1,6 +1,8 @@
 2.5.2 (unreleased)
 ==================
 
+
+* `#949 <https://bitbucket.org/openpyxl/openpyxl/issues/949>`_ High memory use when reading text-heavy files.
 * `#970 <https://bitbucket.org/openpyxl/openpyxl/issues/970>`_ Copying merged cells copies references.
 * `#995 <https://bitbucket.org/openpyxl/openpyxl/issues/995>`_ Cannot delete last row or column.