+++ doc/changes.rst
@@ -6,6 +6,7 @@ Bugfixes
 --------
 
 * `#1017 <https://bitbucket.org/openpyxl/openpyxl/issues/1017>`_ Deleting elements from a legend doesn't work.
+* `#1018 <https://bitbucket.org/openpyxl/openpyxl/issues/1018>`_ Index names repeated for every row in dataframe.
 
 
 2.5.2 (2018-04-06)