+++ doc/changes.rst
@@ -5,6 +5,7 @@
 Bugfixes
 --------
 
+* `#983 <https://bitbucket.org/openpyxl/openpyxl/issues/983>`_ Warning level too aggressive.
 * `#1015 <https://bitbucket.org/openpyxl/openpyxl/issues/1015>`_ Alignment and protection values not saved for named styles.
 * `#1017 <https://bitbucket.org/openpyxl/openpyxl/issues/1017>`_ Deleting elements from a legend doesn't work.
 * `#1018 <https://bitbucket.org/openpyxl/openpyxl/issues/1018>`_ Index names repeated for every row in dataframe.