+++ doc/changes.rst
@@ -10,6 +10,7 @@ Bugfixes
 * `#1017 <https://bitbucket.org/openpyxl/openpyxl/issues/1017>`_ Deleting elements from a legend doesn't work.
 * `#1018 <https://bitbucket.org/openpyxl/openpyxl/issues/1018>`_ Index names repeated for every row in dataframe.
 * `#1020 <https://bitbucket.org/openpyxl/openpyxl/issues/1020>`_ Worksheet protection not being stored.
+* `#1023 <https://bitbucket.org/openpyxl/openpyxl/issues/1023>`_ Exception raised when reading a tooltip.
 
 
 2.5.2 (2018-04-06)